+++ crates/fs/src/ensure_file.rs
@@ -506,322 +506,4 @@ fn strip_dash_suffix(name: &str) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        EnsureFileError, ensure_file, file_equals_bytes, is_transient_rename_error,
-        rename_with_retry, temp_path_for,
-    };
-    use std::{fs, io, path::Path};
-    use tempfile::tempdir;
-
-    #[cfg(unix)]
-    use super::{EMFILE, ENFILE, retry_on_fd_pressure};
-
-    /// New-file path: contents land on disk. Mode handling is covered
-    /// separately in `unix_mode_is_applied_on_new_files`.
-    #[test]
-    fn writes_a_new_file() {
-        let tmp = tempdir().unwrap();
-        let path = tmp.path().join("new.txt");
-
-        ensure_file(&path, b"hello", None).expect("new-file write succeeds");
-
-        assert_eq!(fs::read(&path).unwrap(), b"hello");
-    }
-
-    /// Pre-existing file with matching content short-circuits as
-    /// `Ok(())` and does not touch the target. Mirrors pnpm v11's
-    /// `verifyFileIntegrity(fileDest, integrity) === true` branch in
-    /// `writeBufferToCafs.ts`.
-    #[test]
-    fn existing_target_with_matching_content_is_preserved() {
-        let tmp = tempdir().unwrap();
-        let path = tmp.path().join("existing.txt");
-        fs::write(&path, b"same").unwrap();
-
-        ensure_file(&path, b"same", None).expect("matching contents should short-circuit");
-
-        assert_eq!(fs::read(&path).unwrap(), b"same");
-    }
-
-    /// Pre-existing file with *wrong* contents is a torn-blob case and
-    /// must be atomically replaced with the buffer we were trying to
-    /// write. Mirrors the `writeFileAtomic` branch pnpm takes when
-    /// `verifyFileIntegrity` fails.
-    #[test]
-    fn existing_target_with_wrong_content_is_overwritten_atomically() {
-        let tmp = tempdir().unwrap();
-        let path = tmp.path().join("torn.txt");
-        fs::write(&path, b"garbage-from-crashed-prior-install").unwrap();
-
-        ensure_file(&path, b"fresh", None).expect("torn blob should be rewritten");
-
-        assert_eq!(fs::read(&path).unwrap(), b"fresh");
-        // No leftover temp files in the same directory.
-        let siblings: Vec<_> =
-            fs::read_dir(tmp.path()).unwrap().map(|entry| entry.unwrap().file_name()).collect();
-        assert_eq!(siblings, vec![std::ffi::OsString::from("torn.txt")]);
-    }
-
-    /// Missing parent directory surfaces as a `CreateFile` error
-    /// (kind `NotFound`). Callers are expected to `ensure_parent_dir`
-    /// first; this pins that contract so a regression that quietly
-    /// created ancestors would fail the test.
-    #[test]
-    fn missing_parent_dir_errors() {
-        let tmp = tempdir().unwrap();
-        let path = tmp.path().join("nested/does/not/exist/file.txt");
-
-        let err = ensure_file(&path, b"x", None).expect_err("missing parent should fail");
-        match err {
-            EnsureFileError::CreateFile { error, .. } => {
-                assert_eq!(error.kind(), io::ErrorKind::NotFound);
-            }
-            other => panic!("expected CreateFile/NotFound, got {other:?}"),
-        }
-    }
-
-    /// Unix mode is honoured on the new-file path. Skipped on Windows
-    /// where the `mode` argument is `#[cfg_attr(windows, allow(unused))]`.
-    ///
-    /// Asserts the **owner** bits specifically rather than the full
-    /// `0o777` triplet because `OpenOptionsExt::mode` runs through the
-    /// process umask, which strips group / other bits on systems with
-    /// a restrictive default (e.g. `umask 0o077` CI shells). Owner
-    /// bits are preserved under every sensible umask, so pinning just
-    /// those keeps the test robust without weakening what it verifies
-    /// (that `mode` is being threaded through to the syscall at all
-    /// and that the owner-exec bit survives — the observable property
-    /// that distinguishes an executable CAS blob from a data blob).
-    #[cfg(unix)]
-    #[test]
-    fn unix_mode_is_applied_on_new_files() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let tmp = tempdir().unwrap();
-        let path = tmp.path().join("exec.sh");
-
-        ensure_file(&path, b"#!/bin/sh\n", Some(0o755)).expect("mode-honouring write");
-
-        let mode = fs::metadata(&path).unwrap().permissions().mode() & 0o700;
-        assert_eq!(mode, 0o700, "owner rwx bits of 0o755 must survive any reasonable umask",);
-    }
-
-    /// `-exec` suffix becomes `x` in the temp name (pnpm `removeSuffix`
-    /// parity). Pins the naming scheme so future tweaks stay explicit.
-    #[test]
-    fn temp_path_strips_exec_suffix() {
-        let p = Path::new("/tmp/store/v11/files/ab/cdef-exec");
-        let tmp = temp_path_for(p);
-        let name = tmp.file_name().unwrap().to_string_lossy().into_owned();
-        assert!(name.starts_with("cdefx"), "got {name}");
-    }
-
-    /// Plain hex basenames go through untouched apart from the pid +
-    /// counter suffix.
-    #[test]
-    fn temp_path_passes_plain_basename_through() {
-        let p = Path::new("/tmp/store/v11/files/ab/cdef");
-        let tmp = temp_path_for(p);
-        let name = tmp.file_name().unwrap().to_string_lossy().into_owned();
-        assert!(name.starts_with("cdef"), "got {name}");
-        assert_ne!(name, "cdef", "must include pid + counter suffix");
-    }
-
-    /// Windows AV / indexer interference surfaces as
-    /// `PermissionDenied` or `ResourceBusy` and must trigger the
-    /// retry loop there. On non-Windows those codes are essentially
-    /// always permanent (permission / mount-point issues), so the
-    /// classifier must return `false` to avoid pathologically
-    /// spinning for 60 s on a misconfigured store dir. Any other
-    /// kind must propagate immediately on every platform.
-    #[test]
-    fn transient_rename_error_classifier() {
-        let permission_denied = io::Error::from(io::ErrorKind::PermissionDenied);
-        let resource_busy = io::Error::from(io::ErrorKind::ResourceBusy);
-
-        #[cfg(windows)]
-        {
-            assert!(is_transient_rename_error(&permission_denied));
-            assert!(is_transient_rename_error(&resource_busy));
-        }
-        #[cfg(not(windows))]
-        {
-            assert!(
-                !is_transient_rename_error(&permission_denied),
-                "Unix PermissionDenied is permanent, must not retry",
-            );
-            assert!(
-                !is_transient_rename_error(&resource_busy),
-                "Unix ResourceBusy is effectively permanent, must not retry",
-            );
-        }
-
-        // Non-transient kinds must never trigger the retry loop on
-        // any platform — a regression classifying e.g. `NotFound` as
-        // transient would spin for 60 s on a legitimately missing
-        // source.
-        for kind in [
-            io::ErrorKind::NotFound,
-            io::ErrorKind::AlreadyExists,
-            io::ErrorKind::InvalidInput,
-            io::ErrorKind::InvalidData,
-            io::ErrorKind::Unsupported,
-            io::ErrorKind::Other,
-        ] {
-            assert!(
-                !is_transient_rename_error(&io::Error::from(kind)),
-                "{kind:?} must not be classified as transient"
-            );
-        }
-    }
-
-    /// A symlink at the target path — which on Unix returns `EEXIST`
-    /// from `open(O_CREAT|O_EXCL)` just like a regular file would —
-    /// must be scrubbed and replaced with a real regular file even
-    /// when its target's bytes match what we were about to write.
-    /// Leaving the symlink in place would fool downstream
-    /// `fs::hard_link` (which hardlinks the symlink itself on Linux,
-    /// not the target) and leak non-regular dirents into the CAS.
-    #[cfg(unix)]
-    #[test]
-    fn symlink_at_cas_path_is_scrubbed_to_a_regular_file() {
-        let tmp = tempdir().unwrap();
-        let real_target = tmp.path().join("other_real_file");
-        fs::write(&real_target, b"payload").unwrap();
-
-        let cas_path = tmp.path().join("cas_entry");
-        std::os::unix::fs::symlink(&real_target, &cas_path).unwrap();
-
-        ensure_file(&cas_path, b"payload", None).expect("symlink should be scrubbed");
-
-        let meta = fs::symlink_metadata(&cas_path).unwrap();
-        assert!(
-            meta.file_type().is_file(),
-            "cas_path must be a regular file after scrub, got {:?}",
-            meta.file_type(),
-        );
-        assert_eq!(fs::read(&cas_path).unwrap(), b"payload");
-        // The file the symlink used to point at is untouched — we
-        // replaced the link, not followed it.
-        assert_eq!(fs::read(&real_target).unwrap(), b"payload");
-    }
-
-    /// Dangling symlink (points nowhere) is also scrubbed to a real
-    /// file via the same `symlink_metadata` guard. Without the guard
-    /// we'd still end up in `write_atomic` via the `NotFound` branch
-    /// on `fs::read`, but this pins the expected control flow.
-    #[cfg(unix)]
-    #[test]
-    fn dangling_symlink_at_cas_path_is_scrubbed_to_a_regular_file() {
-        let tmp = tempdir().unwrap();
-        let cas_path = tmp.path().join("cas_entry");
-        std::os::unix::fs::symlink(tmp.path().join("nonexistent"), &cas_path).unwrap();
-
-        ensure_file(&cas_path, b"fresh", None).expect("dangling link should be scrubbed");
-
-        let meta = fs::symlink_metadata(&cas_path).unwrap();
-        assert!(meta.file_type().is_file(), "cas_path must end as a regular file");
-        assert_eq!(fs::read(&cas_path).unwrap(), b"fresh");
-    }
-
-    /// Happy-path rename (no transient errors) moves the payload
-    /// atomically and removes the source. Correctness only — we
-    /// deliberately don't assert a wall-clock bound because rename
-    /// latency on loaded CI / slow filesystems can exceed any
-    /// reasonable timing threshold without the retry path actually
-    /// being taken.
-    #[test]
-    fn rename_with_retry_succeeds_when_no_error() {
-        let tmp = tempdir().unwrap();
-        let src = tmp.path().join("src");
-        let dst = tmp.path().join("dst");
-        fs::write(&src, b"payload").unwrap();
-
-        rename_with_retry(&src, &dst).expect("rename should succeed");
-
-        assert_eq!(fs::read(&dst).unwrap(), b"payload");
-        assert!(!src.exists(), "source should be gone after rename");
-    }
-
-    /// Streaming byte-compare returns `true` iff the file on disk is
-    /// identical to `content`. Pins the three cases
-    /// `verify_or_rewrite` routes through it: exact match (skip
-    /// path), same length but different bytes (atomic rewrite),
-    /// different length (atomic rewrite).
-    #[test]
-    fn file_equals_bytes_classifies_match_mismatch_and_length_mismatch() {
-        let tmp = tempdir().unwrap();
-
-        let equal = tmp.path().join("equal");
-        fs::write(&equal, b"hello world").unwrap();
-        assert!(file_equals_bytes(&equal, b"hello world").unwrap());
-
-        let content_diff = tmp.path().join("content_diff");
-        fs::write(&content_diff, b"hello world").unwrap();
-        assert!(!file_equals_bytes(&content_diff, b"hello WORLD").unwrap());
-
-        // `verify_or_rewrite`'s size-check short-circuits before
-        // reaching this function in practice, but the function
-        // itself still has to classify correctly if called directly.
-        let length_diff = tmp.path().join("length_diff");
-        fs::write(&length_diff, b"short").unwrap();
-        assert!(!file_equals_bytes(&length_diff, b"longer payload").unwrap());
-    }
-
-    /// Multi-chunk files exercise the inner `read_exact` loop rather
-    /// than landing entirely in the first 8 KB read. Guards against
-    /// off-by-one regressions in the chunk-offset math, and confirms
-    /// a byte flipped in the *last* chunk isn't masked by an early
-    /// "first-chunk-matched" short-circuit.
-    #[test]
-    fn file_equals_bytes_handles_multi_chunk_files() {
-        let tmp = tempdir().unwrap();
-        let path = tmp.path().join("big");
-
-        // 20 KB: at least three 8 KB chunks.
-        let content: Vec<u8> = (0..20_000).map(|i| (i % 251) as u8).collect();
-        fs::write(&path, &content).unwrap();
-
-        assert!(file_equals_bytes(&path, &content).unwrap());
-
-        let mut perturbed = content.clone();
-        *perturbed.last_mut().unwrap() ^= 0xff;
-        assert!(!file_equals_bytes(&path, &perturbed).unwrap());
-    }
-
-    /// Transient `EMFILE` / `ENFILE` failures must be retried until
-    /// the underlying op succeeds. Cell-counts attempts so we can
-    /// pin both the predicate ("retries on these errnos") and the
-    /// loop control flow ("returns the first `Ok` value the closure
-    /// produces").
-    #[cfg(unix)]
-    #[test]
-    fn retry_on_fd_pressure_retries_emfile_and_enfile_until_success() {
-        for errno in [EMFILE, ENFILE] {
-            let attempts = std::cell::Cell::new(0);
-            let result = retry_on_fd_pressure(|| {
-                let attempt = attempts.get();
-                attempts.set(attempt + 1);
-                if attempt < 2 { Err(io::Error::from_raw_os_error(errno)) } else { Ok("ok") }
-            });
-            assert_eq!(result.unwrap(), "ok");
-            assert_eq!(attempts.get(), 3, "errno {errno} should have been retried twice");
-        }
-    }
-
-    /// Errors that aren't fd-pressure must propagate immediately —
-    /// retrying would just delay surfacing a real failure (e.g. a
-    /// genuine `NotFound` on the parent dir).
-    #[cfg(unix)]
-    #[test]
-    fn retry_on_fd_pressure_propagates_non_fd_errors() {
-        let attempts = std::cell::Cell::new(0);
-        let result: io::Result<()> = retry_on_fd_pressure(|| {
-            attempts.set(attempts.get() + 1);
-            Err(io::Error::from(io::ErrorKind::NotFound))
-        });
-        assert_eq!(result.unwrap_err().kind(), io::ErrorKind::NotFound);
-        assert_eq!(attempts.get(), 1, "non-fd-pressure errors must not retry");
-    }
-}
+mod tests;
